\name{print.aCGHsegmented}
\alias{print.aCGHsegmented}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for the aCGHsegmented class }
\description{
  Printing the most important feautures of a model returned by \code{jumpoints}.
}
\usage{
\method{print}{aCGHsegmented}(x, digits = max(3, getOption("digits") - 3), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class \code{aCGHsegmented} }
  \item{digits}{ number of digits to be printed }
  \item{\dots}{ arguments passed to other functions }
}
\author{ Vito M.R. Muggeo }

\seealso{\code{\link{jumpoints}}, \code{\link{plot.aCGHsegmented}}}

\keyword{ models }
